requirements.txt
numpy
torch
gymnasium
autorom
stable-baselines3
gym==0.26.2
PyYAML